Move forging base command file to root level directory
