+++ src/base_forging.ts
@@ -15,7 +15,7 @@
 
 import * as inquirer from 'inquirer';
 import { flags as flagParser } from '@oclif/command';
-import BaseIPCCommand from '../../base_ipc';
+import BaseIPCCommand from './base_ipc';
 
 interface Args {
 	readonly address: string;
