+++ src/commands/forging/disable.ts
@@ -13,7 +13,7 @@
  *
  */
 
-import { BaseForgingCommand } from './base';
+import { BaseForgingCommand } from '../../base_forging';
 
 export default class DisableForgingCommand extends BaseForgingCommand {
 	static description = 'Disable forging for given delegate address';
