+++ src/commands/forging/enable.ts
@@ -13,7 +13,7 @@
  *
  */
 
-import { BaseForgingCommand } from './base';
+import { BaseForgingCommand } from '../../base_forging';
 
 export default class EnableForgingCommand extends BaseForgingCommand {
 	static description = 'Enable forging for given delegate address';
